+++ test/integration/migrations_check_roles_enabled/01_first.lua
@@ -1,0 +1,14 @@
+local utils = require('migrator.utils')
+
+return {
+    up = function()
+        if utils.check_roles_enabled({'vshard-router'}) then
+            rawset(_G, 'vshard-router-set', true)
+        end
+
+        if utils.check_roles_enabled({'vshard-storage'}) then
+            rawset(_G, 'vshard-storage-set', true)
+        end
+        return true
+    end
+}