+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- Added `utils.check_roles_enabled` helper function
+  to check whether roles are enabled on the instance (gh-68)
+
 ## [0.6.0]
 ### Added:
 - Configurable timeout for storage migrations (gh-66)